+++ tasks/chyokotov_min_val_by_columns/tests/performance/main.cpp
@@ -11,7 +11,7 @@ class ChyokotovMinValPerfTest : public ppc::util::BaseRunPerfTests<InType, OutTy
   InType input_data_{};
 
   void SetUp() override {
-    const size_t cols = 4000, rows = 4000;
+    const int cols = 4000, rows = 4000;
 
     std::vector<std::vector<int>> matrix(rows, std::vector<int>(cols));
     for (int i = 0; i < rows; i++) {
@@ -35,8 +35,8 @@ class ChyokotovMinValPerfTest : public ppc::util::BaseRunPerfTests<InType, OutTy
 
     std::vector<int> expected_output(matrix[0].size(), INT_MAX);
 
-    for (int i = 0; i < matrix[0].size(); i++) {
-      for (int j = 0; j < matrix.size(); j++) {
+    for (size_t i = 0; i < matrix[0].size(); i++) {
+      for (size_t j = 0; j < matrix.size(); j++) {
         expected_output[i] = std::min(expected_output[i], matrix[j][i]);
       }
     }
